Implemented ~50 clippy suggested (non destructive/disruptive) clarity changes and bumped versions by patch for the next publish to crates.io

+++ src/v2/usizes.rs
@@ -109,7 +109,7 @@ impl EncodingStrategy<usize> for Small {
                 nonzero.encode(writer, &mut ctx.small_nonzero);
                 let value = *value as u64;
                 let zeros = value.leading_zeros() as u8;
-                let bits_beyond_seven = (64 - 7 - zeros) as u8;
+                let bits_beyond_seven = 64 - 7 - zeros ;
                 let bits_beyond_seven: UBits<6> = bits_beyond_seven.try_into().unwrap();
                 bits_beyond_seven.encode(writer, &mut ctx.bits_beyond_seven);
                 for off in 0..6 + u8::from(bits_beyond_seven) {
